Failed attempt on adding a global quiz page showing all the quiz's
Refactored some stuff.
I said failed but it does fetch all the quiz data it's just that i got stuck with adding a media to the quiz itself. it's refusing to work and making a similar attempt to the way i solved question is not working.
quizCard is weird because I cant access my models in it and i tried a lot from @\using, @\inject and MORE
more authorized views , navbar now has the sign in/out at the bottom in desktop (side effect of having the navbar in mobile always open but who tf uses mobile anyway.
also new migrations.